Refactor note update to API controller
